ci: parallelize jobs + add pip-audit + Dependabot (#2)
* ci: parallelize jobs + add pip-audit security + Dependabot

- Split monolithic test job into parallel lint, typecheck, test, security jobs
- Add pip-audit security job with --strict --desc
- Docker build now depends on all 4 jobs passing
- Add Dependabot config for pip, github-actions, and docker ecosystems



* fix: resolve ruff/mypy errors, add pytest-cov, remove duplicate test.yml

- Fix ruff F401 (unused imports) and I001 (import sorting) in cli.py, config.py
- Fix mypy assignment type error in format.py (int -> float parameter)
- Add types-PyYAML, types-psutil to typecheck job for mypy stubs
- Add --ignore-missing-imports to mypy for docker SDK (no stubs available)
- Add pytest-cov to dev dependencies and CI install step
- Remove duplicate test.yml workflow (ci.yml now covers all testing)



* fix: configure ruff to allow BLE001 (resilient monitoring)

---------

+++ src/homelab_monitor/format.py
@@ -8,7 +8,7 @@ from rich.console import Console
 from rich.table import Table
 
 
-def _bytes_to_human(n: int) -> str:
+def _bytes_to_human(n: float) -> str:
     """Convert bytes to a human-readable string."""
     for unit in ("B", "KiB", "MiB", "GiB", "TiB"):
         if abs(n) < 1024.0:
